Experiment with the idea of a driver.rs which "drives" any syncs.
