+++ logins-sql/examples/sync_pass_sql.rs
@@ -32,8 +32,8 @@ use failure::Fail;
 use std::{fs, io::{self, Read, Write}};
 use std::collections::HashMap;
 use fxa_client::{FirefoxAccount, Config, OAuthInfo};
-use sync::{Sync15StorageClientInit, KeyBundle, Store};
-use logins_sql::{PasswordEngine, Login, sync};
+use sync::{Sync15StorageClientInit, KeyBundle, Store, sync_global};
+use logins_sql::{PasswordEngine, LoginsGlobalStateProvider, Login};
 
 const CLIENT_ID: &str = "98adfa37698f255b";
 const REDIRECT_URI: &str = "https://lockbox.firefox.com/fxa/ios-redirect.html";
@@ -454,7 +454,7 @@ fn main() -> Result<()> {
             }
             'S' | 's' => {
                 info!("Syncing!");
-                if let Err(e) = sync(&engine.db, &engine.db, &client_init, &root_sync_key) {
+                if let Err(e) = sync_global(&engine.db, &engine.db, &client_init, &root_sync_key) {
                     warn!("Sync failed! {}", e);
                     warn!("BT: {:?}", e.backtrace());
                 } else {
